+++ src/WidgetRattus/Signal.hs
@@ -44,6 +44,10 @@ module WidgetRattus.Signal
   , zipWith
   , zipWith3
   , zip
+  , parallelWith
+  , parallelWithAwait
+  , parallel
+  , parallelAwait
   , cond
   , update
   , integral
@@ -170,9 +174,9 @@ stop p = jump (box (\ x -> if unbox p x then Just' (const x) else Nothing'))
 -- Example:
 --
 -- >           xs: 1 2 3 4 5   6 7 8   9
--- >           ys:         1 2   3 4 5 6
+-- >           ys:       1 2   3 4 5 6
 -- >
--- > switch xs ys: 1 2 3 1 2 4   3 4 5 6
+-- > switch xs ys: 1 2 3 1 2   3 4 5 6
 switch :: Sig a -> O (Sig a) -> Sig a
 switch (x ::: xs) d = x ::: delay (case select xs d of
                                      Fst   xs'  d'  -> switch xs' d'
@@ -297,6 +301,67 @@ cond = zipWith3 (box (\b x y -> if b then x else y))
 -- > zip = zipWith (box (:*))
 zip :: (Stable a, Stable b) => Sig a -> Sig b -> Sig (a:*b)
 zip = zipWith (box (:*))
+
+
+-- | This is a variant of 'zipWith', but the values passed to the
+-- function may not exist if the corresponding source signal has not
+-- ticked.
+--
+-- Example:
+--
+-- >                            xs:  1            2          3
+-- >                            ys:  1                       0            5
+-- >
+-- > parallelWith (box (:*)) xs ys:  (J 1 :* J 1) (J 2 :* N) (J 3 :* J 0) (N :* J 5)
+
+parallelWith :: Box (Maybe' a -> Maybe' b -> c) -> Sig a -> Sig b -> Sig c
+parallelWith f (x ::: xs) (y ::: ys) =
+   unbox f (Just' x) (Just' y) ::: parallelWithAwait f xs ys
+
+-- | This is a variant of `parallelWith` for delayed signals.
+--
+-- Example:
+--
+-- >                                 xs:    2          3
+-- >                                 ys:               0            5
+-- >
+-- > parallelWithAwait (box (:*)) xs ys:    (J 2 :* N) (J 3 :* J 0) (N :* J 5)
+parallelWithAwait :: Box (Maybe' a -> Maybe' b -> c) -> O (Sig a) -> O (Sig b) -> O (Sig c)
+parallelWithAwait f xs ys = delay (
+  case select xs ys of
+     Fst (x ::: xs')   ys'        -> unbox f (Just' x)  (Nothing') ::: parallelWithAwait f xs' ys'
+     Snd xs'          (y ::: ys') -> unbox f (Nothing') (Just' y)  ::: parallelWithAwait f xs' ys'
+     Both (x ::: xs') (y ::: ys') -> unbox f (Just' x)  (Just' y)  ::: parallelWithAwait f xs' ys')
+
+-- | This is a variant of 'zip', but the signal of pairs only contain
+-- values if the corresponding source signal ticked.
+--
+-- Example:
+--
+-- >             xs:  1            2          3
+-- >             ys:  1                       0            5
+-- >
+-- > parallel xs ys:  (J 1 :* J 1) (J 2 :* N) (J 3 :* J 0) (N :* J 5)
+
+parallel :: Sig a -> Sig b -> Sig (Maybe' a :* Maybe' b)
+parallel (x ::: xs) (y ::: ys) =
+   (Just' x :* Just' y) ::: parallelAwait xs ys
+
+-- | This is a variant of `parallel` for delayed signals.
+--
+-- Example:
+--
+-- >                  xs:    2          3
+-- >                  ys:               0            5
+-- >
+-- > parallelAwait xs ys:    (J 2 :* N) (J 3 :* J 0) (N :* J 5)
+
+parallelAwait :: O (Sig a) -> O (Sig b) -> O (Sig (Maybe' a :* Maybe' b))
+parallelAwait xs ys = delay (
+  case select xs ys of
+     Fst (x ::: xs')   ys'        -> (Just' x  :* Nothing') ::: parallelAwait xs' ys'
+     Snd xs'          (y ::: ys') -> (Nothing' :* Just' y)  ::: parallelAwait xs' ys'
+     Both (x ::: xs') (y ::: ys') -> (Just' x  :* Just' y)  ::: parallelAwait xs' ys')
 
 -- | This function is a variant of 'sample' that works on a delayed
 -- input signal. To this end, 'sampleAwait' takes an additional